move session logic to a react context SessionProvider
