+++ src/pages/settings/village-codes.tsx
@@ -1,6 +1,5 @@
 import { useState } from "react";
 import { trpc } from "@/utils/trpc";
-import { withSession } from "@/lib/session";
 import { VillageCode, NewVillageCode } from "@/db/schema";
 import Link from "next/link";
 
@@ -317,4 +316,4 @@ function VillageCodesPage() {
   );
 }
 
-export default withSession(VillageCodesPage);
+export default VillageCodesPage;
